Testing started

+++ src/Core/Modules/SYSTEM/LogsController.php
@@ -328,7 +328,9 @@ class LogsController extends ModuleInstance {
 		$formatter = new JsonFormatter(JsonFormatter::BATCH_MODE_JSON, true, true);
 		$formatter->includeStacktraces(true);
 		$debugFile = sys_get_temp_dir() . "/{$this->config->main->character}.debug.json";
-		$this->fs->deleteFile($debugFile);
+		if ($this->fs->exists($debugFile)) {
+			$this->fs->deleteFile($debugFile);
+		}
 		$handler = new StreamHandler($debugFile, Logger::DEBUG, true, 0600);
 		$handler->setFormatter($formatter);
 		$processor = new IntrospectionProcessor(Logger::DEBUG, [], 1);
